+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased (????-??-??)
 
-...
+* `<cc-zone-input>`: fix scrollIntoView behaviour
+* `<cc-zone-input>`: center the map
+* `<cc-map>`: fix race conditions on view-zoom update (and other attrs)
 
 ## 5.1.0 (2020-11-03)